+++ CHANGELOG.md
@@ -1,12 +1,12 @@
-# InventoryPlugin (v1.1.0)
+# InventoryPlugin (v1.2.0)
 
 ## Changes
 
-- Enable config autogeneration in `InventoryPlugin` Class
+- Enable update check in `InventoryPlugin` Class
 
 ## Additions
 
-- Added `config.yml` file
+- Add `UpdateManager` Class
 
 ## Deletions